Logo Update (#73)
* Update Logo API

Upload logo was broken (it was not translating image binary properly), fixed those issues. Also added "name" header to get/set logo so company name can be stored alongside logo, and isDefault header to indicate if the returned value is the default or not. If set logo is called without name or logo, the old name or logo will not be overwritten. Updated and added tests for the logo methods.

* Fixed spacing

Fixed some spacing issues

* Tiny formatting change

Removed extra space

* Add Access-Control-Expose-Headers header

Need header Access-Control-Expose-Headers in response in order for cors request to see name and isDefault headers

* Fixed Test Names

Fixed test names and added a null check

* Minor fixes

Added constants and method descriptions